DFK-538 Endra samanlikningsmetode i testar

+++ src/main/kotlin/no/uutilsynet/testlab2testing/aggregering/AggregertResultatSide.kt
@@ -9,7 +9,7 @@ data class AggregertResultatSide(
     val loeysing: Loeysing,
     val sideUrl: URL,
     val sideNivaa: Int,
-    val gjennomsnittligBruddProsentTR: Float?,
+    val gjennomsnittligBruddProsentTR: Double?,
     val talElementSamsvar: Int,
     val talElementBrot: Int,
     val talElementVarsel: Int,
