+++ src/main/kotlin/no/uutilsynet/testlab2testing/aggregering/AggregertResultatTestregel.kt
@@ -16,6 +16,6 @@ data class AggregertResultatTestregel(
     val talSiderSamsvar: Int,
     val talSiderBrot: Int,
     val talSiderIkkjeForekomst: Int,
-    val testregelGjennomsnittlegSideSamsvarProsent: Float?,
-    val testregelGjennomsnittlegSideBrotProsent: Float?,
+    val testregelGjennomsnittlegSideSamsvarProsent: Double?,
+    val testregelGjennomsnittlegSideBrotProsent: Double?,
 ) : AutotesterTestresultat
